Add files via upload
Graduate Admission Prediction Using Machine Learning

This project aims to predict the chances of admission into graduate programs based on various factors. The dataset used in this project includes several parameters which are considered important during the application for Masters Programs. The parameters included are:
GRE Scores (out of 340)
TOEFL Scores (out of 120)
University Rating (out of 5)
Statement of Purpose (SOP) Strength (out of 5)
Letter of Recommendation (LOR) Strength (out of 5)
Undergraduate GPA (out of 10)
Research Experience (either 0 or 1)
The model built in this project uses these parameters to predict the probability of a student being admitted to a graduate program.

Please note that this is a machine learning model and the predictions are based on the data it was trained on. Therefore, actual results may vary. Always consult with a professional when making important educational decisions.